Refactor some design elements from the editor into their own components.
